[gc] Make `GCRootRef` nullable
Main motivation for this change is for the purpose of implementing the JNI. In the JNI a nullptr (aka null root) is used to represent null references as the JNI does not have a concept of "roots" vs "references/pointers" (`jobjects` are always roots).
In most cases differentiating between a null root and a root containing a null reference is not required which is why the boolean conversion treats them identically.
Since a null root may not be assigned an object, the `operator=` taking `T*` was given the explicit name `assign` to reduce the chance of accidentally doing so.
Since in the vast majority of cases one assigned a `GCRootRef` to another `GCRootRef` I am not anticipating this to be a problem.