[interpreter] Reuse single interpreter entry in `JIT2InterpreterLayer` (#300)
This layer has so far generated unnecessarily large amounts of LLVM IR
by needlessly compiling a new interpreter entry each time.

This PR simplifies its implementation by reusing the single interpreter
entry created previously. The layer is now only responsible for
translating from the JIT CC to the interpreter CC. This sadly still has
to happen per-method rather than per-method-type (as in the
interpreter2jit translation) as we require method objects to be
prepended, but JIT -> Interpreter calls are expected not to be hot
anyway.