[unwind] Speed up installing OSR entry (#313)
These are two minor micro optimizations that speed up unwinding and
installing the OSR Entry.
The first makes sure that any user of our libunwind has a flag set that
only unwinding state for the native architecture has to be included.
This affects the size of an `UnwindFrame` reducing it from ca. 1.2kB to
roughly 300 bytes.
Second, the dynamic memory allocation of the `ForcedException` has been
removed in favor of using a thread local static, removing a call to
`new`.

---------

Co-authored-by: nemegyeb <nemegyeb@gmail.com>